+++ extensions/bmrks/src/components/authenticated-view.tsx
