+++ extensions/bmrks/src/search-bookmarks.tsx
@@ -5,7 +5,7 @@ import { useGroups } from "../lib/use-groups";
 import { useBookmarks } from "../lib/use-bookmarks";
 import { getHostname } from "../lib/get-hostname";
 import { deleteBookmark, moveBookmarkToGroup } from "../lib/db";
-import AuthenticatedView from "./components/AuthenticatedView";
+import AuthenticatedView from "./components/authenticated-view";
 
 export function SearchBookmarks({ user }: { user: User }) {
   const { data: groups, isLoading: isLoadingGroups } = useGroups(user);
